Customization screen bug fixed
Although the screen does not errors anymore, there is still a strange
bug. Females will randomly have missing hair and/or brows, and the
corresponding ui buttons to change the piece type will also be missing.

The game will crash as soon as a bastioneer is spawned in the world, no
idea why.